feat(adapters): extend centralized metrics requirements system to RAGAS and Opik adapters
- Add metric test case creation for Opik adapter with support for:
  - OpikSpanTestCase for span-level metrics
  - OpikSessionTestCase for session-level metrics
  - OpikHallucinationTestCase for specialized hallucination detection
- Add metric test case creation for RAGAS adapter with:
  - RagasMultiTurnTestCase for multi-turn conversation analysis
- Add metric configuration files for both adapters
- Update adapter implementations to use centralized metrics requirements
- Add comprehensive test file for RAGAS and Opik metrics
